backend/requirements.txt
absl-py==2.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.128.0
fastdtw==0.3.4
flatbuffers==25.2.10
fonttools==4.58.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
jax==0.6.1
jaxlib==0.6.1
joblib==1.5.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numba==0.61.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
starlette==0.50.0
tdqm==0.0.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
tslearn==0.6.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.40.0
websockets==15.0.1
